requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
jinja2==3.1.2
aiohttp==3.8.5 